nhb-apps/requirements/requirements_dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements_dev.in
#
arrow==0.17.0             # via -r requirements_test.txt, djangosaml2idp
asgiref==3.3.1            # via -r requirements_test.txt, django
beautifulsoup4==4.9.3     # via -r requirements_dev.in
cachetools==4.1.1         # via -r requirements_test.txt, google-auth
certifi==2020.11.8        # via -r requirements_test.txt, requests
cffi==1.14.4              # via -r requirements_test.txt, cryptography
chardet==3.0.4            # via -r requirements_test.txt, requests
coverage==5.3             # via -r requirements_dev.in
cryptography==3.2.1       # via -r requirements_test.txt, pyopenssl, pysaml2
defusedxml==0.6.0         # via -r requirements_test.txt, pysaml2
django-debug-toolbar==3.1.1  # via -r requirements_dev.in
django-extensions==3.1.0  # via -r requirements_dev.in
django==3.0.11            # via -r requirements_test.txt, django-debug-toolbar, djangosaml2idp
djangosaml2idp==0.7.2     # via -r requirements_test.txt
google-api-core==1.23.0   # via -r requirements_test.txt, google-api-python-client
google-api-python-client==1.12.8  # via -r requirements_test.txt
google-auth-httplib2==0.0.4  # via -r requirements_test.txt, google-api-python-client
google-auth==1.23.0       # via -r requirements_test.txt, google-api-core, google-api-python-client, google-auth-httplib2
googleapis-common-protos==1.52.0  # via -r requirements_test.txt, google-api-core
httplib2==0.18.1          # via -r requirements_test.txt, google-api-python-client, google-auth-httplib2
idna==2.10                # via -r requirements_test.txt, requests
packaging==20.5           # via pip-review
pip-licenses==3.1.0       # via -r requirements_dev.in
pip-review==1.1.0         # via -r requirements_dev.in
protobuf==3.14.0          # via -r requirements_test.txt, google-api-core, googleapis-common-protos
psycopg2-binary==2.8.6    # via -r requirements_test.txt
ptable==0.9.2             # via pip-licenses
pyasn1-modules==0.2.8     # via -r requirements_test.txt, google-auth
pyasn1==0.4.8             # via -r requirements_test.txt, pyasn1-modules, rsa
pycparser==2.20           # via -r requirements_test.txt, cffi
pyopenssl==20.0.0         # via -r requirements_test.txt, pysaml2
pyotp==2.4.1              # via -r requirements_test.txt
pyparsing==2.4.7          # via packaging
pysaml2==6.3.1            # via -r requirements_test.txt, djangosaml2idp
python-dateutil==2.8.1    # via -r requirements_test.txt, arrow, pysaml2
pytz==2020.4              # via -r requirements_test.txt, django, djangosaml2idp, google-api-core, pysaml2
qrcode==6.1               # via -r requirements_test.txt
requests==2.25.0          # via -r requirements_test.txt, google-api-core, pysaml2
rsa==4.6                  # via -r requirements_test.txt, google-auth
six==1.15.0               # via -r requirements_test.txt, cryptography, google-api-core, google-api-python-client, google-auth, google-auth-httplib2, protobuf, pyopenssl, pysaml2, python-dateutil, qrcode
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.4.1           # via -r requirements_test.txt, django, django-debug-toolbar
uritemplate==3.0.1        # via -r requirements_test.txt, google-api-python-client
urllib3==1.26.2           # via -r requirements_test.txt, requests

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
